Refactor spdx state
