--- frontend/src/context/spdx/index.ts
@@ -1,2 +1,0 @@
-export * from './spdx.context';
-export * from './spdx.types';
